Crate cplwm_api

The API you must implement

This crate contains the definitions of the various traits you will implement and some basic types. You are not allowed to edit this crate.

wm contains the various traits that you must implement in the assignments.

types contains the basic types used throughout the whole project.

Modules

types
Basic types used throughout the whole project.

wm
Window manager functionality defined as traits