Struct cplwm_api::types::WindowLayout

pub struct WindowLayout {
    pub focused_window: Option<Window>,
    pub windows: Vec<(Window, Geometry)>,
}

As explained in the assignment, the WindowLayout struct fully describes the layout of the windows that the window manager should display.

This data structure is passed to the backend to make sure the display server actually displays the desired window layout of the window manager.

Fields
focused_window: Option<Window>

The focused window or None in which case no window is focused.

Invariant: if focused_window = Some(w), w must be in windows.

windows: Vec<(Window, Geometry)>

A Vec of all the visible windows along with their geometry.

The order of this Vec determines the stacking order of the windows: the first window is the bottom window, the last window is the top window, see the ASCII figure below.

+-------+
| last  |-----+
+-------+     |
      | first |
      |       |
      +-------+

This main reason for choosing this order is that it is easier to append an element to the end of a vector than it is to prepend one.

Methods

impl WindowLayout

fn new() -> WindowLayout

Construct a new, empty window layout without any windows and with nothing focused.