Struct cplwm_api::types::WindowWithInfo

pub struct WindowWithInfo {
    pub window: Window,
    pub geometry: Geometry,
    pub float_or_tile: FloatOrTile,
    pub fullscreen: bool,
}

A WindowWithInfo is the combination of a Window with additional information: its Geometry, whether it should float or not (float_or_tile), and whether it should be displayed fullscreen or not (fullscreen).

The float_or_tile field will be set to Float by the backend when the window is a dialog or popup, otherwise Tile. Window managers not implementing FloatSupport can safely ignore this.

Analogously for the fullscreen field, it will be set to true by the backend when the window wants to be displayed fullscreen, otherwise false. Window managers not implementing FullscreenSupport can safely ignore this.

This is a separate type used by the add_window and get_window_info methods of the WindowManager trait, and will also be useful when defining a window manager data type yourself.

Fields
window: Window

The window.

geometry: Geometry

The geometry of the window.

float_or_tile: FloatOrTile

Indicate whether the window should float or tile.

fullscreen: bool

Indicate whether the window should be displayed fullscreen or not.

Methods

impl WindowWithInfo

fn new_tiled(window: Window, geometry: Geometry) -> WindowWithInfo

Create a new tiled non-fullscreen WindowWithInfo.

fn new_float(window: Window, geometry: Geometry) -> WindowWithInfo

Create a new floating non-fullscreen WindowWithInfo.

fn new_fullscreen(window: Window, geometry: Geometry) -> WindowWithInfo

Create a new fullscreen WindowWithInfo.

The float_or_tile field is set to Tile.

fn new(window: Window, geometry: Geometry, float_or_tile: FloatOrTile, fullscreen: bool) -> WindowWithInfo

Create a new WindowWithInfo with the given arguments.