Module cplwm_api::wm

Window manager functionality defined as traits

This module defines a number of traits, each corresponding to an assignment. Every window manager must at least implement the WindowManager trait.

Interaction Between Traits

When writing a window manager that implements multiple traits, e.g., TilingSupport and FloatSupport, take care of the interaction between the traits. For example, when calling swap_with_master (from the TilingSupport trait) with a window that is currently floating (FloatSupport), a possible thing to do is to tile the window first and then swap it with the master window. Or when calling toggle_floating with a minimised window (MinimiseSupport), unminimise it first before floating it. Try to make reasonable and consistent choices when it comes to these edge cases (keep the listed invariants in mind). Document them and write test cases for them.

Traits

FloatSupport
A window manager that supports floating windows.

FullscreenSupport
A window manager that supports fullscreen windows.

GapSupport
A window manager that supports gaps between tiles.

MinimiseSupport
A window manager that supports (un)minimising windows.

MultiWorkspaceSupport
A window manager that has multiple workspaces.

TilingSupport
A window manager that supports tiling.

WindowManager
A basic window manager.