Trait cplwm_api::wm::FloatSupport

pub trait FloatSupport: WindowManager {
    fn get_floating_windows(&self) -> Vec<Window>;
    fn toggle_floating(&mut self, window: Window) -> Result<(), Self::Error>;
    fn set_window_geometry(&mut self, window: Window, new_geometry: Geometry) -> Result<(), Self::Error>;

    fn is_floating(&self, window: Window) -> bool { ... }
}

A window manager that supports floating windows.

Some windows are not suitable for tiling, e.g. dialogs, popups, video, etc. Instead of tiling them, they should float above the tiled windows. Floating windows can be moved around and resized with the mouse.

The backend will detect that some windows should float from the start and will set the float_or_tile field of WindowWithInfo to Float in these cases. A window manager implementing this trait should implement the add_window method of WindowManager such that these windows will float after adding them.

The user can choose to float any tiled window or to sink (make it tiled again) any floating window.

Invariant: every floating window w (is_floating(w) == true) must be placed above the tiled windows in the window layout returned by get_window_layout().

Required Methods

fn get_floating_windows(&self) -> Vec<Window>

Return a vector of all the visible floating windows.

The order of the windows in the vector does not matter.

fn toggle_floating(&mut self, window: Window) -> Result<(), Self::Error>

If the given window is floating, let it sink, if it is not floating, let it float.

When a non-floating window starts to float, its original geometry (passed to add_window) should be restored.

Invariant: if calling toggle_floating(w) with a tiled window w succeeds, is_floating(w) must return true.

Invariant: if calling toggle_floating(w) with a floating window w succeeds, is_floating(w) must return false.

Invariant: the result of is_floating(w) must be the same before and after calling toggle_floating(w) twice.

This function is allowed to return an appropriate error when the window is not managed by the window manager.

fn set_window_geometry(&mut self, window: Window, new_geometry: Geometry) -> Result<(), Self::Error>

Resize/move the given floating window according to the given geometry.

This function is called when the user moves or resizes a window using the mouse, but can also be called by custom user commands.

The window layout should reflect the geometry change of the floating window.

This function is allowed to return an appropriate error when the window is not managed by the window manager or when the window is not floating.

Provided Methods

fn is_floating(&self, window: Window) -> bool

Return true if the given window is floating.

This function must always return false when the given window is not managed by the window manager.

Invariant: if is_floating(w) == true for some window w, then is_managed(w) == true.

Invariant: is_floating(w) == true for some window w, iff the vector returned by the get_floating_windows method contains w.

A default implementation is provided in terms of get_floating_windows(). Override this implementation if you have a more efficient one.