Trait cplwm_api::wm::FullscreenSupport

pub trait FullscreenSupport: WindowManager {
    fn get_fullscreen_window(&self) -> Option<Window>;
    fn toggle_fullscreen(&mut self, window: Window) -> Result<(), Self::Error>;
}

A window manager that supports fullscreen windows.

Users wishing to watch a video fullscreen, to play a game fullscreen, or to view any application fullscreen, can make any window become fullscreen using this trait.

There can at most one window be fullscreen at a time.

The backend will detect that some windows want to be fullscreen from the start and will set the fullscreen field of WindowWithInfo to true in these cases. A window manager implementing this trait should implement the add_window method of WindowManager such that these windows will be fullscreen after adding them.

A user can make any window fullscreen by using the toggle_fullscreen.

Think carefully about the interaction between a fullscreen window and the other traits: when an action is performed on a fullscreen window, should it stop being fullscreen? What when an action is performed on another window which is not the fullscreen one? In some cases it might make sense to keep the window fullscreen (e.g., when another window is removed), in some cases maybe not (e.g., when a non-fullscreen window is added). Try to make reasonable and consistent choices (keep the invariants of this and the other traits in mind). Document the choices and write test cases for them.

Required Methods

fn get_fullscreen_window(&self) -> Option<Window>

Return the current fullscreen, if any.

Invariant: if get_fullscreen_window() == Some(w), then is_managed(w) == true.

Invariant: if get_fullscreen_window() == Some(w), then get_focused_window() == Some(w).

fn toggle_fullscreen(&mut self, window: Window) -> Result<(), Self::Error>

Make the given window fullscreen, or when it is already fullscreen, undo it.

When called on a window that is already fullscreen, it should restore the window to the state before, e.g. float at the same place. Hint: you could use the float_or_tile field of WindowWithInfo.

Invariant: if calling toggle_fullscreen(w) with a window w that is not yet fullscreen, w should be the only visible window according to get_window_layout, its geometry should be the same size as the screen, and get_fullscreen_window(w) == Some(w).

The window layout before and after calling toggle_fullscreen twice with the currently focused should be the same. This cannot hold for a window manager that implements TilingSupport. Try to figure out why.