Trait cplwm_api::wm::GapSupport

pub trait GapSupport: WindowManager {
    fn get_gap(&self) -> GapSize;
    fn set_gap(&mut self, GapSize);
}

A window manager that supports gaps between tiles.

The user can configure the gap size at run-time. The gaps are only shown between tiled windows. Floating windows and the fullscreen window are unaffected. Initially the gap is 0.

For example the user has the following tiles:

+----------+----------+
|          |          |
|          |          |
|    1     |    2     |
|          |          |
|          |          |
+----------+----------+

When the gap is set to 5, it is as if every window gets an invisible 5 pixel border of empty space. Windows don't share the gap in between them. Every tiled window is moved 5 pixels down and to 5 pixels to the right. The width and height of every tiled window is also shrunk by 10 pixels.

+---------------------+
|+--------+ +--------+|
||        | |        ||
||    1   | |   2    ||
||        | |        ||
|+--------+ +--------+|
+---------------------+

Even when there is only a single tile there should be a gap around it.

You may ignore scenarios in which the gap size is so large that one of the windows might become invisible.

Implementors of this trait must adapt their implementation of the get_window_layout() method accordingly.

Required Methods

fn get_gap(&self) -> GapSize

Return the current gap size.

Initially 0.

fn set_gap(&mut self, GapSize)

Set the gap size.

Invariant: after setting set_gap(g) with some gap size g, get_gap() == g.