Trait cplwm_api::wm::MinimiseSupport

pub trait MinimiseSupport: WindowManager {
    fn get_minimised_windows(&self) -> Vec<Window>;
    fn toggle_minimised(&mut self, window: Window) -> Result<(), Self::Error>;

    fn is_minimised(&self, window: Window) -> bool { ... }
}

A window manager that supports (un)minimising windows.

Remember that a tiling window manager displays each window in a different tile. So when you have opened ten applications, the screen will be split in ten tiles. Sometimes you want to open an application without having to look at it the whole time (MultiWorkspaceSupport is also useful for this scenario). In that case, minimising the window is the solution: you can temporarily hide the window and regain the screen space without having to close the window.

Naturally, there is also a way to reveal these windows again by unminimising them.

Note: methods of other traits like focus_window, focus_window, toggle_fullscreen, ... called with a minimised window as argument should first unminimise the window.

Hint: you can use remove_window and add_window to hide and reveal windows.

Required Methods

fn get_minimised_windows(&self) -> Vec<Window>

Return a vector of all the minimised windows.

The order of the windows in the vector does matter.

The windows must occur in the order they were minimised: the window that was minimised first must occur first in the vector, the window that was minimised last must occur last. This makes it easy to define a function that unminimises the last minimised window.

fn toggle_minimised(&mut self, window: Window) -> Result<(), Self::Error>

Minimise the given window, or when it is already minimised, unminise it.

When a minimised floating window is unminimised, it should float again and have the same geometry as before. Hint: you could use the float_or_tile field of WindowWithInfo. Analogously for fullscreen windows.

Invariant: if calling toggle_minimised(w) with an unminimised window w succeeds, w may no longer be visible according to get_window_layout and is_minimised(w) must return true.

Invariant: if calling toggle_minimised(w) with an already minimised window w succeeds, w must be visible according to get_window_layout and is_minimised(w) must return false.

The window layout before and after minimising and directly unminimising the currently focused window should be the same. This cannot hold for a window manager that implements TilingSupport. Try to figure out why.

Provided Methods

fn is_minimised(&self, window: Window) -> bool

Return true if the given window is minimised.

This function must always return false when the given window is not managed by the window manager.

Invariant: if is_minimised(w) == true for some window w, then is_managed(w) == true.

Invariant: is_minimised(w) == true for some window w, iff the vector returned by the get_minised_windows method contains w.

A default implementation is provided in terms of get_minimised_windows(). Override this implementation if you have a more efficient one.