Trait cplwm_api::wm::MultiWorkspaceSupport

pub trait MultiWorkspaceSupport<WM: WindowManager>: WindowManager {
    fn get_current_workspace_index(&self) -> WorkspaceIndex;
    fn get_workspace(&self, index: WorkspaceIndex) -> Result<&WM, Self::Error>;
    fn get_workspace_mut(&mut self, index: WorkspaceIndex) -> Result<&mut WM, Self::Error>;
    fn switch_workspace(&mut self, index: WorkspaceIndex) -> Result<(), Self::Error>;
}

A window manager that has multiple workspaces.

For a small introduction to workspaces, see the first section of the assignment.

The idea is that there are a number of different workspaces (MAX_WORKSPACE_INDEX + 1), each represented by a different copy of a single type of window manager. Initially the first (index 0) workspace is active. The user interacts with this workspace's window manager, e.g., adding/removing windows, focusing a window, etc. When the user then switches to another workspace, all the windows will be hidden. The user can then interact with this workspace as if it is a new one. When the user switches back to the first workspace, the windows previously opened in this workspace are shown again. When the user switches back to the other workspace, the windows previously opened in this workspace are shown again.

This is useful when multitasking: one workspace is for surfing the web, another for working on a CPL assignment, a third one for working on your thesis, etc.

Think carefully about the interaction between this trait and the others. Most actions must only be executed on the current workspace, e.g., add_window, but some must be executed on all workspaces, e.g., set_gap or resize_screen. Even the getters are non-obvious: should get_windows return the windows of all workspaces or only the current one? What about get_floating_windows? Try to make reasonable and consistent choices (keep the invariants of this and the other traits in mind). Document the choices and write test cases for them.

Required Methods

fn get_current_workspace_index(&self) -> WorkspaceIndex

Return the current workspace index.

When creating a new workspace this will return 0.

Invariant: 0 <= get_current_workspace_index() <= MAX_WORKSPACE_INDEX.

fn get_workspace(&self, index: WorkspaceIndex) -> Result<&WM, Self::Error>

Get an immutable borrow of the workspace at the given index.

This function should return an appropriate error when 0 <= index <= MAX_WORKSPACE_INDEX is not true.

fn get_workspace_mut(&mut self, index: WorkspaceIndex) -> Result<&mut WM, Self::Error>

Get a mutable borrow of the workspace at the given index.

This function should return an appropriate error when 0 <= index <= MAX_WORKSPACE_INDEX is not true.

fn switch_workspace(&mut self, index: WorkspaceIndex) -> Result<(), Self::Error>

Switch to the workspace at the given index.

If index == get_current_workspace_index(), do nothing.

Invariant: the window layout after switching to another workspace and then switching back to the original workspace should be the same as before.

This function should return an appropriate error when 0 <= index <= MAX_WORKSPACE_INDEX is not true.