Trait cplwm_api::wm::TilingSupport

pub trait TilingSupport: WindowManager {
    fn get_master_window(&self) -> Option<Window>;
    fn swap_with_master(&mut self, window: Window) -> Result<(), Self::Error>;
    fn swap_windows(&mut self, dir: PrevOrNext);
}

A window manager that supports tiling.

The idea of tiling is explained in the assignment. While most tiling window managers have many possible layouts to (dynamically) choose from, we limit ourselves to one simple layout here. In an optional assignment you will be asked to implement a different layout algorithm.

Layout algorithm

This window manager follows the following the tiling layout: we start out with no windows. When the first window is added, the window is maximised as in the ASCII diagram below.

+---------------------+
|                     |
|                     |
|          1          |
|                     |
|                     |
+---------------------+

When a second window is added, the screen is split in two tiles: a left tile for window 1 and a right tile for window 2.

+----------+----------+
|          |          |
|          |          |
|    1     |    2     |
|          |          |
|          |          |
+----------+----------+

When a third window is added, the right tile will split in two tiles: a top tile for window 2 and a bottom tile for window 3.

+---------------------+
|          |          |
|          |    2     |
|    1     +----------+
|          |          |
|          |    3     |
+----------+----------+

The left tile will never be split, we call this the master tile. The user typically places his/her main application in this tile, e.g., the browser or editor. Additional windows, e.g. a terminal, a file manager, or a chat window are displayed in the side tiles. Note that even when the master tile is focused, new windows will not appear in the master tile, but in a new side tile.

When a fourth window is added, an additional tile is created on the right side and the new window is placed in the bottom tile.

+----------+----------+
|          |    2     |
|          +----------+
|    1     |    3     |
|          +----------+
|          |    4     |
+----------+----------+

When window 2, 3, or 4 is closed, the corresponding tile disappears and we go back to the previous layout. When window 1 is closed, the first side window (2) is chosen to be displayed in the master tile. The promoted window's previous tile disappears.

Invariant: at all times there must be as many tiles as there are visible windows. Note that this will not hold when there are floating (FloatSupport) or minimised windows (MinimiseSupport).

How to implement it

This trait contains some useful methods to move around windows. It is not sufficient to implement these methods, the actual tiling logic happens when windows are added or removed, and in get_window_layout. Consequently, you must change the appropriate implementations of the methods of the WindowManager trait to take care of tiling.

Required Methods

fn get_master_window(&self) -> Option<Window>

Return the window displayed in the master tile.

If there are no windows, return None.

Invariant: get_master_window() == Some(w), then w must occur in the vector returned by get_windows().

Invariant: if the vector returned by get_windows() is empty => get_master_window() == None. The other direction of the arrow must not hold, e.g., there could floating windows (see FloatSupport), but no tiled windows.

fn swap_with_master(&mut self, window: Window) -> Result<(), Self::Error>

Swap the given window with the window in the master tile.

After the function has succeeded, the master window should be focused.

If the given window is already in the master tile, no windows have to be swapped, but the master window should be focused.

Invariant: if swap_with_master(w) succeeds, get_master_window() == Some(w).

This function is allowed to return an appropriate error when the window is not managed by the window manager.

fn swap_windows(&mut self, dir: PrevOrNext)

Swap the focused window with the one in the next or previous tile.

Do nothing when there are no windows, when there is only one window, or when no window is focused.

If there were two tiles and the swap happened, the same window will be focused, but the other tile will be focused.

Invariant: calling swap_windows(dir) for any dir will not change the focused window, even if no window was focused.

Invariant: calling swap_windows(dir) and then swap_windows(dir.opposite()) will not change the window layout.