Trait cplwm_api::wm::WindowManager

pub trait WindowManager: Encodable + Decodable + Debug + Clone {
    type Error: Error + 'static;
    fn new(screen: Screen) -> Self;
    fn get_window_layout(&self) -> WindowLayout;
    fn get_windows(&self) -> Vec<Window>;
    fn add_window(&mut self, window_with_info: WindowWithInfo) -> Result<(), Self::Error>;
    fn remove_window(&mut self, window: Window) -> Result<(), Self::Error>;
    fn focus_window(&mut self, window: Option<Window>) -> Result<(), Self::Error>;
    fn cycle_focus(&mut self, dir: PrevOrNext);
    fn get_window_info(&self, window: Window) -> Result<WindowWithInfo, Self::Error>;
    fn get_screen(&self) -> Screen;
    fn resize_screen(&mut self, screen: Screen);

    fn is_managed(&self, window: Window) -> bool { ... }
    fn get_focused_window(&self) -> Option<Window> { ... }
}

A basic window manager.

Every window manager defined by a struct or enum representing its state, must implement this trait.

Supertraits

Note what comes after WindowManager:, Encodable + Decodable + Debug + Clone. These are the “supertraits” of WindowManager: to implement this trait, these four other traits must be implemented as well.

Encodable and Decodable

The first two, Encodable and Decodable, indicate that it should be possible to serialise and deserialise a window manager. This functionality is used in the backends, which allow reloading a live window manager without losing its state. This is useful during development: one can easily switch to a newly compiled version of the window manager without losing all the state. Before reloading, the whole window manager is serialised to JSON which is written to /tmp. The window manager is then restarted from scratch. During startup, the window manager tries to read this JSON, deserialises it and uses the result as the window manager, including all the state instead of a new empty window manager. Windows that have been added or removed since the state was saved will be added and removed to/from the restored window manager.

All this is taken care of in the backend, so the only thing you have to do is make sure your window manager is Encodable and Decodable. Lucky for you, this traits can be derived automatically using the derive annotation.

For example:

#[derive(RustcEncodable, RustcDecodable, ...)]
pub struct MyWM {
    ...
}

In the example, a struct MyWM with some fields is defined. The derive annotation on the struct will implement both interfaces automatically for you. In order for this to work, every field of the struct must implement these interfaces too. Most basic types available in Rust (i32, Vec, ...) already implement this, but if you define additional structs, you also have to put the same annotation on them.

Note that you can also implement these traits manually, but this will be a lot more work.

Debug

The Debug trait is used to convert a window manager to a string representation to print it during development, in test output or when tracing.

This trait can also be derived automatically using #[derive(Debug)].

Clone

The Clone is used to make a clone of a window manager.

This trait can also be derived automatically using #[derive(Clone)].

In Summary

The example below demonstrates that you can implement these four traits using one line of code:

#[derive(RustcEncodable, RustcDecodable, Debug, Clone)]
pub struct MyWM {
    ...
}

You also have to make sure that every type you use in the definition of your window manager struct (or enum) also implements these four traits.

Associated Types

type Error: Error + 'static

The type of the errors of the window manager.

You can define your own enum with different types of errors you window manager can return. Many functions of this trait will return a Result<(), Self::Error>: either an Ok(()), meaning that everything went ok, or an Err(err) where err is of this Error type, meaning that an error occurred.

Note the two “supertraits” your error type must implement: error::Error and 'static. The first is the base trait for all Rust errors, click on it to see the methods you must implement. This trait also has three “supertraits”:

Debug: a toString-like method for debugging that can be derived automatically using #[derive(Debug, ...)].

Display: a toString-like method that will be used to show error to users (as opposed to developers). This method has to be manually implemented.

Reflect: you can safely ignore this.

The second trait is not an actual trait, but a constraint that says that your error must be static and should not contain any references with a limited lifetime. In other words, your error type may not contain any type &'a T.

Examples

/// Error type for my window manager
#[derive(Debug)]  // Implement `Debug` automatically
pub enum MyWMError {
    /// This window is not known by the window.
    UnknownWindow(Window),
    // Feel free to add more
    ...
}
// Manually implement the `Display` trait.
use std::fmt;
impl fmt::Display for MyWMError {
    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
        match *self {
            MyWMError::UnknownWindow(ref window) =>
                write!(f, "Unknown window: {}", window),
            ...
        }
    }
}
// Now implement the `Error` trait.
use std::error;
impl error::Error for MyWMError {
    fn description(&self) -> &'static str {
        match *self {
            // Because we must return a `&'static str`, we can't
            // include the window in it, because that creates a
            // `String`.
            MyWMError::UnknownWindow(_) => "Unknown window",
        }
    }
}
// We can now use `MyWMError` as the `Error` type of the `MyWM` window
// manager.
impl WindowManager for MyWM {
    type Error = MyWMError;
    ...
}

Note that you can use the same error type in multiple window managers.

Required Methods

fn new(screen: Screen) -> Self

The constructor of a window manager.

The sole argument is the Screen, useful to determine the size of fullscreen windows or to center a window.

fn get_window_layout(&self) -> WindowLayout

Return the current desired window layout.

For more information about this, see the assignment and the documentation for WindowLayout.

fn get_windows(&self) -> Vec<Window>

Return a vector of all the windows managed by the window manager, visible or not.

The order of the windows in the vector does not matter.

Invariant: get_windows() must not contain duplicates.

fn add_window(&mut self, window_with_info: WindowWithInfo) -> Result<(), Self::Error>

Add a new window along with its information, e.g. its Geometry.

This is called whenever a new window is created.

Invariant: is_managed must return true for the given window after add_window was called with the given window.

Invariant: after adding a window using add_window, it must be focused according to get_focused_window.

A window manager that implements FloatSupport should float the window when the float_or_tile field of WindowWithInfo is set to Float.

A window manager that implements FullscreenSupport should make the window fullscreen when the fullscreen field of WindowWithInfo is set to true.

This function is allowed to return an appropriate error when the window is already managed by the window manager.

fn remove_window(&mut self, window: Window) -> Result<(), Self::Error>

Remove the given window from the window manager.

This is called whenever a window is destroyed/killed.

The same window remains focused, unless the focused window has been removed.

Invariant: is_managed must return false for the given window after remove_window was called with the given window.

This function is allowed to return an appropriate error when the window is not managed by the window manager.

fn focus_window(&mut self, window: Option<Window>) -> Result<(), Self::Error>

Focus the given window, or when passed None, focus nothing.

This is called when the user clicks or hovers on(to) a window, or changes the focus using the keyboard.

Invariant: when focus_window succeeds, get_focused_window must return the same argument.

This function is allowed to return an appropriate error when the window is not managed by the window manager.

fn cycle_focus(&mut self, dir: PrevOrNext)

Focus the previous or next window.

Do nothing when there are no windows. When there is only one window, focus it if currently no window is focused, otherwise do nothing.

When no window is focused, any window may become focused.

Cycling the focus back and forth shouldn't change the focused window.

fn get_window_info(&self, window: Window) -> Result<WindowWithInfo, Self::Error>

Get the info (WindowWithInfo) belonging to the given window. It should reflect the current state (location/size, floating or tiled, fullscreen or not) of the window.

This method will come in handy when implementing later assignments by defining wrappers around existing window managers.

This function should return an appropriate error when the window is not managed by the window manager.

fn get_screen(&self) -> Screen

Return the screen managed by the window manager.

fn resize_screen(&mut self, screen: Screen)

Resize the screen according to the given Screen.

This is called whenever the resolution of the screen is changed.

Note that we do not support multiple monitors.

Invariant: after resize_screen is called with a screen, get_screen() must return the same screen.

Provided Methods

fn is_managed(&self, window: Window) -> bool

Check whether a window is managed by this window manager.

Invariant: is_managed(w) == true for some window w iff the vector returned by the get_windows method contains w.

A default implementation is provided in terms of get_windows. Override this implementation if you have a more efficient one.

fn get_focused_window(&self) -> Option<Window>

Return the window that is currently focused according to the window manager.

If no window should be focused, return None.

Invariant: get_focused_window() == get_window_layout().focused_window.

Invariant: get_focused_window() == Some(w) then is_managed(w) == true.

A default implementation is provided in terms of get_window_layout and its focused_window field. Override this implementation if you have a more efficient one.